Add curl easy socket backend: consolidate mqttcurl into mqttnet.
